+++ src/Warehouses/BrewUp.Warehouse.SharedKernel/DomainIds/Stock.cs
@@ -1,3 +1,3 @@
 ﻿namespace BrewUp.Warehouse.SharedKernel.DomainIds;
 
-public record Stock(double Value);
+public record Stock(decimal Value);
